+++ NEWS.md
@@ -1,5 +1,6 @@
 # odeqIRtools 0.2.0
 * Add strd_lyr_display to crate file to pass to Dave Pray. 
+* Better error reporting for Al_default_DOC
 
 # odeqIRtools 0.1.0