+++ NEWS.md
@@ -1,3 +1,6 @@
+# odeqIRtools 0.3.0
+* Add archive_IRdb() to archive previous cycle IR database
+
 # odeqIRtools 0.2.0
 * Add strd_lyr_display to crate file to pass to Dave Pray. 
 * Better error reporting for Al_default_DOC